refactor: moved Validator classes to Controllers
