+++ Uni_life_tracker/app/src/main/java/com/generic/plannr/Controllers/InputTextValidator.java
@@ -1,4 +1,4 @@
-package com.generic.plannr;
+package com.generic.plannr.Controllers;
 
 import android.graphics.Color;
 import android.widget.TextView;
